Refactor Profile model and UpdateProfileMutation
This commit refactors the Profile model in the users/models.py file by removing the "image_hash" field and modifying the "has_required_fields" property to return a list of required fields that are not filled.

In addition, the UpdateProfileMutation in the users/mutations.py file now returns the list of required fields as "required_fields" instead of a boolean value.

The ProfileType in the users/types.py file now includes a "required_fields" field that returns the list of required fields. The "has_required_fields" property has been modified to check if the "required_fields" list is not empty.